add ConsensusStrategy abstraction
state_transition is higher level code while casper_utils contains more
specific consensus strategy, so all casper logic should live in
casper_utils.

With ConsensusStrategy casper_utils/state_transition circular
dependency problem is fixed.
